# docs/swagger.yaml
basePath: /api/v2
definitions:
  apperrors.ErrorResponse:
    properties:
      error_response:
        type: string
    type: object
...
  doc_structs.BoardResponse:
    properties:
      board:
        $ref: '#/definitions/entities.Board'
    type: object
...
  doc_structs.ThumbnailUploadResponse:
    properties:
      thumbnail_url:
        $ref: '#/definitions/dto.UrlObj'
    type: object
...
  dto.BoardID:
    properties:
      board_id:
        type: integer
    type: object
...
  dto.NewBoardInfo:
    properties:
      description:
        type: string
      name:
        type: string
      owner_email:
        type: string
      owner_id:
        type: integer
      workspace_id:
        type: integer
    type: object
...
  dto.UpdatedBoardInfo:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  dto.UpdatedBoardThumbnailInfo:
    properties:
      id:
        type: integer
      thumbnail:
        items:
          type: integer
        type: array
    type: object
...
  dto.UrlObj:
    properties:
      url:
        type: string
    type: object
...
  dto.UserInfo:
    properties:
      email:
        type: string
      id:
        type: integer
    type: object
...
  dto.UserPublicInfo:
    properties:
      avatar_url:
        type: string
      description:
        type: string
      email:
        type: string
      name:
        type: string
      surname:
        type: string
      user_id:
        type: integer
    type: object
...
  entities.Board:
    properties:
      board_id:
        type: integer
      description:
        type: string
      lists:
        items:
          $ref: '#/definitions/entities.List'
        type: array
      name:
        type: string
      owner:
        $ref: '#/definitions/dto.UserInfo'
      thumbnail_url:
        type: string
      users:
        items:
          $ref: '#/definitions/dto.UserPublicInfo'
        type: array
    type: object
  entities.List:
    properties:
      board_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      list_position:
        type: integer
      name:
        type: string
      tasks:
        items:
          $ref: '#/definitions/entities.Task'
        type: array
    type: object
  entities.Task:
    properties:
      date_created:
        type: string
      description:
        type: string
      end:
        type: string
      id:
        type: integer
      list_id:
        type: integer
      list_position:
        type: integer
      name:
        type: string
      start:
        type: string
      users:
        items:
          $ref: '#/definitions/entities.User'
        type: array
    type: object
  entities.User:
    properties:
      avatar_url:
        type: string
      description:
        type: string
      email:
        type: string
      name:
        type: string
      password_hash:
        type: string
      surname:
        type: string
      user_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: kdanil01@mail.ru
    name: Капитанов Даниил
    url: https://vk.com/poophead27
  description: Лучшее и единственно приложение, имитирующее Trello.
  license:
    name: None
    url: None
  title: LA TABULA API
  version: "2.0"
paths:
...
  /board/:
    post:
      consumes:
      - application/json
      description: Получить доску
      parameters:
      - description: id доски
        in: body
        name: boardID
        required: true
        schema:
          $ref: '#/definitions/dto.BoardID'
      produces:
      - application/json
      responses:
        "200":
          description: объект доски
          schema:
            $ref: '#/definitions/doc_structs.BoardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Получить доску
      tags:
      - boards
  /board/create/:
    post:
      consumes:
      - application/json
      description: Создать доску
      parameters:
      - description: данные новой доски
        in: body
        name: newBoardInfo
        required: true
        schema:
          $ref: '#/definitions/dto.NewBoardInfo'
      produces:
      - application/json
      responses:
        "200":
          description: объект доски
          schema:
            $ref: '#/definitions/doc_structs.BoardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Создать доску
      tags:
      - boards
  /board/delete/:
    delete:
      consumes:
      - application/json
      description: Удалить доску
      parameters:
      - description: id доски
        in: body
        name: boardID
        required: true
        schema:
          $ref: '#/definitions/dto.BoardID'
      produces:
      - application/json
      responses:
        "204":
          description: no content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Удалить доску
      tags:
      - boards
  /board/update/:
    post:
      consumes:
      - application/json
      description: Обновить доску
      parameters:
      - description: обновленные данные доски
        in: body
        name: boardInfo
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatedBoardInfo'
      produces:
      - application/json
      responses:
        "204":
          description: no content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Обновить доску
      tags:
      - boards
  /board/update/change_thumbnail/:
    post:
      consumes:
      - application/json
      description: Обновить картинку доску
      parameters:
      - description: обновленные данные задания
        in: body
        name: boardInfo
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatedBoardThumbnailInfo'
      produces:
      - application/json
      responses:
        "200":
          description: Ссылка на новую картинку
          schema:
            $ref: '#/definitions/doc_structs.ThumbnailUploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Обновить картинку доски
      tags:
      - boards
...
swagger: "2.0"